refactor: subway에서 auth 분리
