+++ src/main/java/nextstep/subway/member/application/MemberService.java
@@ -2,7 +2,7 @@ package nextstep.subway.member.application;
 
 import nextstep.subway.member.application.dto.MemberRequest;
 import nextstep.subway.member.application.dto.MemberResponse;
-import nextstep.subway.member.client.dto.ProfileResponse;
+import nextstep.auth.client.dto.ProfileResponse;
 import nextstep.subway.member.domain.LoginMember;
 import nextstep.subway.member.domain.Member;
 import nextstep.subway.member.domain.MemberRepository;
