+++ src/main/java/nextstep/subway/member/ui/MemberController.java
@@ -1,5 +1,6 @@
 package nextstep.subway.member.ui;
 
+import nextstep.auth.ui.AuthenticationPrincipal;
 import nextstep.subway.member.application.MemberService;
 import nextstep.subway.member.application.dto.MemberRequest;
 import nextstep.subway.member.application.dto.MemberResponse;
